refactor: sbom_manager
