+++ cve_bin_tool/cli.py
@@ -70,8 +70,8 @@ from cve_bin_tool.log import LOGGER
 from cve_bin_tool.merge import MergeReports
 from cve_bin_tool.output_engine import OutputEngine
 from cve_bin_tool.package_list_parser import PackageListParser
-from cve_bin_tool.sbom_detection import sbom_detection
-from cve_bin_tool.sbom_manager import SBOMManager
+from cve_bin_tool.sbom_manager.parse import SBOMParse
+from cve_bin_tool.sbom_manager.sbom_detection import sbom_detection
 from cve_bin_tool.util import ProductInfo
 from cve_bin_tool.version import VERSION
 from cve_bin_tool.version_scanner import VersionScanner
@@ -1048,13 +1048,13 @@ def main(argv=None):
         if args["sbom_file"]:
             sbom_root = args["sbom_file"]
             # Process SBOM file
-            sbom_list = SBOMManager(
+            sbom_list = SBOMParse(
                 args["sbom_file"],
                 sbom_type=args["sbom"],
                 logger=LOGGER,
                 validate=not args["disable_validation_check"],
             )
-            parsed_data = sbom_list.scan_file()
+            parsed_data = sbom_list.parse_sbom()
             LOGGER.info(
                 f"The number of products to process from SBOM - {len(parsed_data)}"
             )
