+++ fuzz/fuzz_cyclonedx.py
@@ -17,7 +17,7 @@ from google.protobuf.json_format import MessageToDict
 import fuzz.generated.cyclonedx_pb2 as cyclonedx_pb2
 
 with atheris.instrument_imports():
-    from cve_bin_tool.sbom_manager import SBOMManager
+    from cve_bin_tool.sbom_manager.parse import SBOMParse
 
 
 def TestParseData(data):
@@ -50,8 +50,8 @@ def TestParseData(data):
         with open(file_path, "w") as f:
             json.dump(json_data, f)
 
-        sbom_engine = SBOMManager(file_path, sbom_type="cyclonedx")
-        sbom_engine.scan_file()
+        sbom_engine = SBOMParse(file_path, sbom_type="cyclonedx")
+        sbom_engine.parse_sbom()
 
     except SystemExit:
         return
