+++ test/test_output_engine.py
@@ -20,6 +20,7 @@ from rich.console import Console
 from cve_bin_tool.output_engine import OutputEngine, output_csv, output_json, output_pdf
 from cve_bin_tool.output_engine.console import output_console
 from cve_bin_tool.output_engine.util import format_output
+from cve_bin_tool.sbom_manager.generate import SBOMGenerate
 from cve_bin_tool.util import (
     CVE,
     CVEData,
@@ -1150,19 +1151,23 @@ class TestOutputEngine(unittest.TestCase):
         self.mock_file = tempfile.NamedTemporaryFile("w+", encoding="utf-8")
 
     def test_generate_sbom(self):
+        """Test SBOM generation"""
         with patch(
-            "cve_bin_tool.output_engine.SBOMPackage"
-        ) as mock_sbom_package, patch("cve_bin_tool.output_engine.SBOMRelationship"):
+            "cve_bin_tool.sbom_manager.generate.SBOMPackage"
+        ) as mock_sbom_package, patch(
+            "cve_bin_tool.sbom_manager.generate.SBOMRelationship"
+        ):
             mock_package_instance = MagicMock()
             mock_sbom_package.return_value = mock_package_instance
 
-            self.output_engine.generate_sbom(
+            sbomgen = SBOMGenerate(
                 all_product_data=self.all_product_data,
                 filename="test.sbom",
                 sbom_type="spdx",
                 sbom_format="tag",
                 sbom_root="CVE-SCAN",
             )
+            sbomgen.generate_sbom()
 
             # Assertions
             mock_package_instance.set_name.assert_any_call("CVEBINTOOL-CVE-SCAN")
@@ -1206,9 +1211,7 @@ class TestOutputEngine(unittest.TestCase):
                 mock_package_instance.get_package.return_value,
                 mock_package_instance.get_package.return_value,
             }
-            actual_packages = [
-                package for package in self.output_engine.sbom_packages.values()
-            ]
+            actual_packages = [package for package in sbomgen.sbom_packages.values()]
             self.assertEqual(actual_packages, list(expected_packages))
 
     def tearDown(self) -> None:
