+++ test/test_sbom.py
@@ -7,9 +7,9 @@ from pathlib import Path
 import pytest
 
 from cve_bin_tool.input_engine import TriageData
-from cve_bin_tool.sbom_detection import sbom_detection
-from cve_bin_tool.sbom_manager import Remarks, SBOMManager
-from cve_bin_tool.util import ProductInfo
+from cve_bin_tool.sbom_manager.parse import SBOMParse
+from cve_bin_tool.sbom_manager.sbom_detection import sbom_detection
+from cve_bin_tool.util import ProductInfo, Remarks
 
 
 class TestSBOM:
@@ -117,8 +117,8 @@ class TestSBOM:
         (str(SBOM_PATH / "nonexistent.spdx.json"),),
     )
     def test_nonexistent_file(self, filepath: str):
-        sbom_engine = SBOMManager(filepath)
-        assert sbom_engine.scan_file() == {}
+        sbom_engine = SBOMParse(filepath)
+        assert sbom_engine.parse_sbom() == {}
 
     @pytest.mark.parametrize(
         "filename, sbom_type",
@@ -129,8 +129,8 @@ class TestSBOM:
         ),
     )
     def test_invalid_file(self, filename: str, sbom_type: str):
-        sbom_engine = SBOMManager(filename, sbom_type)
-        assert sbom_engine.scan_file() == {}
+        sbom_engine = SBOMParse(filename, sbom_type)
+        assert sbom_engine.parse_sbom() == {}
 
     @pytest.mark.parametrize(
         "filename, sbom_type",
@@ -140,8 +140,8 @@ class TestSBOM:
         ),
     )
     def test_invalid_type(self, filename: str, sbom_type: str):
-        sbom_engine = SBOMManager(filename, sbom_type)
-        assert sbom_engine.scan_file() == {}
+        sbom_engine = SBOMParse(filename, sbom_type)
+        assert sbom_engine.parse_sbom() == {}
 
     @pytest.mark.parametrize(
         "filename, spdx_parsed_data",
@@ -158,8 +158,8 @@ class TestSBOM:
     def test_valid_spdx_file(
         self, filename: str, spdx_parsed_data: dict[ProductInfo, TriageData]
     ):
-        sbom_engine = SBOMManager(filename, sbom_type="spdx")
-        scan_result = sbom_engine.scan_file()
+        sbom_engine = SBOMParse(filename, sbom_type="spdx")
+        scan_result = sbom_engine.parse_sbom()
         for p in spdx_parsed_data:
             assert p in scan_result
 
@@ -175,8 +175,8 @@ class TestSBOM:
     def test_valid_cyclonedx_file(
         self, filename: str, cyclonedx_parsed_data: dict[ProductInfo, TriageData]
     ):
-        sbom_engine = SBOMManager(filename, sbom_type="cyclonedx")
-        scan_result = sbom_engine.scan_file()
+        sbom_engine = SBOMParse(filename, sbom_type="cyclonedx")
+        scan_result = sbom_engine.parse_sbom()
         for p in cyclonedx_parsed_data:
             assert p in scan_result
 
@@ -191,8 +191,8 @@ class TestSBOM:
     def test_bad_ext_ref_cyclonedx_file(
         self, filename: str, cyclonedx_parsed_data: dict[ProductInfo, TriageData]
     ):
-        sbom_engine = SBOMManager(filename, sbom_type="cyclonedx")
-        scan_result = sbom_engine.scan_file()
+        sbom_engine = SBOMParse(filename, sbom_type="cyclonedx")
+        scan_result = sbom_engine.parse_sbom()
         for p in cyclonedx_parsed_data:
             assert p in scan_result.keys()
 
@@ -208,8 +208,8 @@ class TestSBOM:
     def test_ext_ref_priority_cyclonedx_file(
         self, filename: str, cyclonedx_parsed_data: dict[ProductInfo, TriageData]
     ):
-        sbom_engine = SBOMManager(filename, sbom_type="cyclonedx")
-        scan_result = sbom_engine.scan_file()
+        sbom_engine = SBOMParse(filename, sbom_type="cyclonedx")
+        scan_result = sbom_engine.parse_sbom()
         for p in cyclonedx_parsed_data:
             assert p in scan_result.keys()
 
@@ -220,8 +220,8 @@ class TestSBOM:
     def test_valid_swid_file(
         self, filename: str, swid_parsed_data: dict[ProductInfo, TriageData]
     ):
-        sbom_engine = SBOMManager(filename, sbom_type="swid")
-        scan_result = sbom_engine.scan_file()
+        sbom_engine = SBOMParse(filename, sbom_type="swid")
+        scan_result = sbom_engine.parse_sbom()
         for p in swid_parsed_data:
             assert p in scan_result
 
@@ -235,8 +235,8 @@ class TestSBOM:
     def test_common_prefix_split(self, product, version, productinfo, no_existent_file):
         """Unit Test for common_prefix_split that try to split on hyphen if no vendors are
         are found and the product has hyphen, here a no_existent_file is used
-        with sole purpose for creating a SBOMManager instance"""
-        sbom_engine = SBOMManager(no_existent_file)
+        with sole purpose for creating a SBOMParse instance"""
+        sbom_engine = SBOMParse(no_existent_file)
         scanned_list = sbom_engine.common_prefix_split(product, version)
         assert productinfo in scanned_list
 
@@ -263,8 +263,8 @@ class TestSBOM:
         if file does not match schema or if xml data is parsed against wrong type of sbom
         (indicated by validate being set to False)
         """
-        sbom_engine = SBOMManager(filename, sbom_type, validate=validate)
-        assert sbom_engine.scan_file() == {}
+        sbom_engine = SBOMParse(filename, sbom_type, validate=validate)
+        assert sbom_engine.parse_sbom() == {}
 
     @pytest.mark.parametrize(
         "filename, expected_sbom_type",
